feat: accompany 타입 분리
