+++ src/components/AccompanyList/AccompanyListItem.tsx
@@ -1,14 +1,7 @@
 import styled from "@emotion/styled";
 import { Text } from "@/styles/Text";
 import { useNavigate } from "react-router-dom";
-
-export interface AccompanyListItemProps {
-    id: number;
-    title: string;
-    content: string;
-    timeAgo: string;
-    imageUrl?: string;
-}
+import { AccompanyListItemProps } from "@/types/accompany";
 
 export const AccompanyListItem = ({ id, title, content, timeAgo, imageUrl }: AccompanyListItemProps) => {
     const navigate = useNavigate();
