+++ src/types/accompany.ts
@@ -1,0 +1,7 @@
+export interface AccompanyListItemProps {
+    id: number;
+    title: string;
+    content: string;
+    timeAgo: string;
+    imageUrl?: string;
+}
